# Conflicts:
#	src/pline.c
#	sys/winnt/windmain.c
#	win/tty/topl.c
#	win/win32/scripts/fuzzer/longtest.bat
#	win/win32/scripts/fuzzer/restore.bat
#	win/win32/scripts/fuzzer/runtill.bat

+++ src/sp_lev.c
@@ -3762,7 +3762,7 @@ struct opvar *mc;
                     break;
                 case 0:
                 case 1:
-                    if (levl[x][y].lit == (unsigned int) lit)
+                    if (levl[x][y].lit == lit)
                         selection_setpoint(x, y, ret, 1);
                     break;
                 }
